Renamed vector-2d to vector_2d to be compliant with code standard
