+++ test/test_main.cpp
@@ -1,5 +1,5 @@
 #define CATCH_CONFIG_MAIN // This tells Catch to provide a main() - only do this in one cpp file
-#include "../src/vector-2d.cpp"
+#include "../src/vector_2d.cpp"
 
 #include "../src/Image.cpp"
